requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.9.2
numpy #==2.1.1
tensorflow==2.17.0
pyzmq==26.0.3
pydantic-settings
uvicorn
aioredis==1.3.1
redis
aiohttp
Pillow